app/requirements.txt
Django>=2.1.3,<2.2.0
django-crispy-forms==1.8.1
xhtml2pdf 
django-qr-code